+++ src/CardInfoForm.jsx
@@ -1,7 +1,9 @@
 import styled from "styled-components";
 
-export const CardInfoForm = styled.form`
+const CardInfoForm = styled.form`
   display: flex;
   flex-direction: column;
   gap: 20px;
 `;
+
+export default CardInfoForm;
