+++ src/components/CardPreview.jsx
