+++ src/components/index.jsx
@@ -1,0 +1,6 @@
+export { default as CardPreview } from "./CardPreview";
+export { default as CardNumberInput } from "./CardNumberInput";
+export { default as CardExpireDateInput } from "./CardExpireDateInput";
+export { default as CardHolderNameInput } from "./CardHolderNameInput";
+export { default as CardSecurityCodeInput } from "./CardSecurityCodeInput";
+export { default as CardPasswordInput } from "./CardPasswordInput";
